pyproject.toml
[tool.poetry]
name = "ats-toolkit"
version = "2.0.0"
description = "Modular Cybersecurity Framework - 144 Tools"
authors = ["Eric Dopi <eric@dopiiii.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/Dopiiii/ATS-Toolkit"
keywords = ["cybersecurity", "osint", "pentest", "red-team", "security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

[tool.poetry.dependencies]
python = "^3.11"

# Core
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"

# Async
aiohttp = "^3.9.0"
httpx = "^0.25.0"

# DNS
dnspython = "^2.4.0"

# Logging
structlog = "^23.2.0"

# Data
pyyaml = "^6.0.1"
pandas = "^2.1.0"
sqlalchemy = "^2.0.0"

# Security
cryptography = "^41.0.0"

# TUI (Phase 1)
textual = "^0.38.1"
rich = "^13.7.0"

# Web UI (Phase 2)
streamlit = "^1.28.0"

# Redis (optional)
redis = {version = "^5.0.0", optional = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.0"
ruff = "^0.1.0"
pre-commit = "^3.6.0"

[tool.poetry.extras]
redis = ["redis"]
full = ["redis"]

[tool.poetry.scripts]
ats = "main:main"
ats-api = "src.api.main:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501"]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
